refactor: Update project structure
Moved contents of backend folder to main project folder for better project structure
