+++ .gitignore
@@ -1,3 +1,8 @@
 .idea
 /Test Results - Unittests_in_test_pms_py.html
 __pycache__
+venv
+.env
+app.py
+test_database.db
+database.db
