+++ requirements.txt
@@ -3,7 +3,7 @@ click==7.1.2
 Flask==1.1.2
 Flask-Migrate==2.5.3
 Flask-Script==2.0.6
-Flask-SQLAlchemy==2.4.3
+Flask-SQLAlchemy==2.4.4
 itsdangerous==1.1.0
 Jinja2==2.11.2
 Mako==1.1.3
